solve bug when adding foods to meals. Complete food addition function
