+++ lib/foods.js
@@ -27,12 +27,12 @@ const orderById = (a, b) => {
  }
 
 
-$('#food-list').on('click', '.delete-food-button', function (event) {
-  event.preventDefault()
-  event.target.parentElement.parentElement.remove()
-  getMeals(event.target.id)
-  deleteCall(event.target.id)
-})
+// $('#food-list').on('click', '.delete-food-button', function (event) {
+//   event.preventDefault()
+//   event.target.parentElement.parentElement.remove()
+//   getMeals(event.target.id)
+//   deleteCall(event.target.id)
+// })
 
 $('#food-list').on('click', '.food-name-field', function (event){
   let tableCell = $(`<td class="food-name-field" id='${event.currentTarget.id}'><input type="text" name="food-name" value="${event.target.textContent}"
@@ -60,33 +60,33 @@ $('#food-list').on('click', '.food-calorie-field', function (event){
   });
 })
 
-function deleteCall (id) {
-  return fetch(`https://vast-retreat-17218.herokuapp.com/api/v1/foods/${id}`, {
-    method: 'DELETE'
-  }).catch(error => console.error({ error }))
-}
+// function deleteCall (id) {
+//   return fetch(`https://vast-retreat-17218.herokuapp.com/api/v1/foods/${id}`, {
+//     method: 'DELETE'
+//   }).catch(error => console.error({ error }))
+// }
 
-const getMeals = foodId => {
-  fetch('https://vast-retreat-17218.herokuapp.com/api/v1/meals')
-    .then(response => response.json())
-    .then(parsedResponse => {
-      parsedResponse.forEach(function (element) {
-        element.foods.forEach(function (food) {
-          // this is an array
-          if (food.id == foodId) {
-            deleteFoodFromMeals(element.id, food.id)
-          }
-        })
-      })
-    })
-}
+// const getMeals = foodId => {
+//   fetch('https://vast-retreat-17218.herokuapp.com/api/v1/meals')
+//     .then(response => response.json())
+//     .then(parsedResponse => {
+//       parsedResponse.forEach(function (element) {
+//         element.foods.forEach(function (food) {
+//           // this is an array
+//           if (food.id == foodId) {
+//             deleteFoodFromMeals(element.id, food.id)
+//           }
+//         })
+//       })
+//     })
+// }
 
-const deleteFoodFromMeals = (mealID, foodId) => {
-  fetch(
-    `https://vast-retreat-17218.herokuapp.com/api/v1/meals/${mealID}/foods/${foodId}`,
-    { method: 'DELETE' }
-  ).catch(error => console.error({ error }))
-}
+// const deleteFoodFromMeals = (mealID, foodId) => {
+//   fetch(
+//     `https://vast-retreat-17218.herokuapp.com/api/v1/meals/${mealID}/foods/${foodId}`,
+//     { method: 'DELETE' }
+//   ).catch(error => console.error({ error }))
+// }
